requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.29
django-adminplus==0.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
gunicorn==19.6.0
idna==2.7
psycopg2==2.9.9
python-decouple==3.1
pytz==2018.5
requests==2.19.1
urllib3==1.23
whitenoise==3.3.1
asgiref==3.8.1
certifi==2018.8.24
chardet==3.0.4
charset-normalizer==3.4.2
dj-database-url==0.5.0
Django==5.0
django-adminplus==0.5
django-rest-framework==0.1.0
djangorestframework==3.16.0
gunicorn==19.6.0
idna==2.7
psycopg==3.2.9
python-decouple==3.1
pytz==2018.5
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==5.0.1
